+++ dflib/src/test/java/com/nhl/dflib/DataFrame_GetColumnTest.java
@@ -12,7 +12,7 @@ public class DataFrame_GetColumnTest {
 
     @Test
     public void testGetColumn_byLabel() {
-        DataFrame df = DataFrame.newFrame("a", "b").foldByRow(
+        DataFrame df = DataFrame.foldByRow("a", "b").of(
                 1, "x",
                 2, "y");
 
@@ -23,7 +23,7 @@ public class DataFrame_GetColumnTest {
 
     @Test
     public void testGetColumn_byPosition() {
-        DataFrame df = DataFrame.newFrame("a", "b").foldByRow(
+        DataFrame df = DataFrame.foldByRow("a", "b").of(
                 1, "x",
                 2, "y");
 
@@ -32,6 +32,7 @@ public class DataFrame_GetColumnTest {
         new SeriesAsserts(cb).expectData(1, 2);
     }
 
+    @Deprecated
     @Test
     public void testGetColumnAsInt_byLabel() {
         DataFrame df = DataFrame
@@ -43,78 +44,84 @@ public class DataFrame_GetColumnTest {
         new IntSeriesAsserts(cb).expectData(3, 6, -1);
     }
 
+    @Deprecated
     @Test
     public void testGetColumnAsInt_byPosition() {
         DataFrame df = DataFrame
-                .newFrame("a", "b")
-                .columns(Series.of("a", "b", "x"), Series.ofInt(3, 6, -1));
+                .byColumn("a", "b")
+                .of(Series.of("a", "b", "x"), Series.ofInt(3, 6, -1));
 
         IntSeries cb = df.getColumnAsInt(1);
 
         new IntSeriesAsserts(cb).expectData(3, 6, -1);
     }
 
+    @Deprecated
     @Test
     public void testGetColumnAsInt_NotIntLabel() {
         DataFrame df = DataFrame
-                .newFrame("a", "b")
-                .columns(Series.of("a", "b", "x"), Series.ofInt(3, 6, -1));
+                .byColumn("a", "b")
+                .of(Series.of("a", "b", "x"), Series.ofInt(3, 6, -1));
 
         assertThrows(IllegalArgumentException.class, () -> df.getColumnAsInt("a"));
     }
 
 
+    @Deprecated
     @Test
     public void testGetColumnAsDouble_byLabel() {
         DataFrame df = DataFrame
-                .newFrame("a", "b")
-                .columns(Series.of("a", "b", "x"), Series.ofDouble(3., 6.3, -1.01));
+                .byColumn("a", "b")
+                .of(Series.of("a", "b", "x"), Series.ofDouble(3., 6.3, -1.01));
 
         DoubleSeries cb = df.getColumnAsDouble("b");
 
         new DoubleSeriesAsserts(cb).expectData(3., 6.3, -1.01);
     }
 
+    @Deprecated
     @Test
     public void testGetColumnAsDouble_byPosition() {
         DataFrame df = DataFrame
-                .newFrame("a", "b")
-                .columns(Series.of("a", "b", "x"), Series.ofDouble(3., 6.3, -1.01));
+                .byColumn("a", "b")
+                .of(Series.of("a", "b", "x"), Series.ofDouble(3., 6.3, -1.01));
 
         DoubleSeries cb = df.getColumnAsDouble(1);
 
         new DoubleSeriesAsserts(cb).expectData(3., 6.3, -1.01);
     }
 
+    @Deprecated
     @Test
     public void testGetColumnAsDouble_NotDoubleLabel() {
         DataFrame df = DataFrame
-                .newFrame("a", "b")
-                .columns(Series.of("a", "b", "x"), Series.ofDouble(3., 6.3, -1.01));
+                .byColumn("a", "b")
+                .of(Series.of("a", "b", "x"), Series.ofDouble(3., 6.3, -1.01));
 
         assertThrows(IllegalArgumentException.class, () -> df.getColumnAsDouble("a"));
     }
 
+    @Deprecated
     @Test
     public void testGetColumnAsLong_byLabel() {
         DataFrame df = DataFrame
-                .newFrame("a", "b")
-                .columns(Series.of("a", "b", "x"), Series.ofLong(3, 6, -1));
+                .byColumn("a", "b")
+                .of(Series.of("a", "b", "x"), Series.ofLong(3, 6, -1));
 
         LongSeries cb = df.getColumnAsLong("b");
 
         new LongSeriesAsserts(cb).expectData(3, 6, -1);
     }
 
+    @Deprecated
     @Test
     public void testGetColumnAsLong_byPosition() {
         DataFrame df = DataFrame
-                .newFrame("a", "b")
-                .columns(Series.of("a", "b", "x"), Series.ofLong(3, 6, -1));
+                .byColumn("a", "b")
+                .of(Series.of("a", "b", "x"), Series.ofLong(3, 6, -1));
 
         LongSeries cb = df.getColumnAsLong(1);
 
         new LongSeriesAsserts(cb).expectData(3, 6, -1);
     }
-
 }